TINKERPOP-1504 Added the property key to events for new properties.
Newly added properties triggered "changed" events when using EventStrategy but the specific key that was newly added was not available in the mutation listener. The listener would simply report the new value only. After this change the listener now gets a DetachedVertexPropery/Property with the newly added key and a null value set.