TINKERPOP-1653 Allow multiple scripts to be passed to gremlin.sh
This is a non-breaking change and simply allows multiple -i and -e options on gremlin.sh. The old method for passing a single script name to gremlin.sh remains intact as well. Had to write a custom parser for the command line arguments to do this because it doesn't appear that CliBuilder is capable of parsing the in the manner necessary to support this feature unfortunately.