Fixed a bug in how global bindings were tracked/applied.
There were several problems here (which ultimately affected the 3.3.0 branch), but mainly global bindings were being re-applied to the ScriptContext given to eval() and thus making them both engine and global bindings which then led to local variables getting assigned to the global scope. Also, getGlobalBindings() on GremlinExecutor is no longer necessary and is deprecated as the global bindings should be manipulated directly on the GremlinScriptEngine. CTR