TINKERPOP-1608 Fix for xsltproc processing
From comments by Robert Dale on the PR itself: xsltproc is complaining because the apply-templates is not immediately after a new node when trying to apply selected attributes. From what I can tell, the only attributes that should be in the <edge> element are already explicitly defined. It is not necessary to copy attributes (@*) back into the <edge> element. That appears to be the main difference between TP2 and TP3 - that the label attribute is now a <data> element with key attribute. Thus, the attribute selection can be removed.