TINKERPOP-1427 Fixed problems with v2 vs v3 in gremlin-python
There were several problems. In protocol.py there was an expectation as the to the format of the ResponseMessage that changed between v2 and v3. I added a bit of a sketchy hack to deal with that detects v2/v3 and then parses accordingly. That should be nicer. There were some issues with tests as well that bound assertions to v2 so that when the default serializer swapped from v2 to v3 those assertions started failing. I forced those specific tests to v2 to get them to pass. Ultimately, we need to more generally test v2 and v3, but at least gremlin-python is defaulted to v3 at this point and all tests are passing.