Used WeakReference for evaluationFuture
Adjusted a previous commit - didn't get the last one right for some reason. The evaluationFuture holds on to the result and thus needs the WeakReference. In this way the result can be garbage collected once it goes out of scope and doesn't need to wait for the timeout to expire. CTR