Initialization scripts to gremlin server won't timeout
Prior to this change the timeout for init scripts was the same as the the timeout configured for the server. Since init scripts might need to be long run (e.g. a long load script to init a graph into memory) it would force the user to make their timeout setting really high. A really high setting might be really bad though for basic request/response. Perhaps in the future it would be better to make this a configuration but for now this change should suffice. CTR