Disabled the jython build
The jython plugin is failing because of a recent change to pypi. The plugin needs an updated for it to start working again. Discussed here:

https://lists.apache.org/thread.html/178ef75505155d9a795a5f119722dcd33f83f5c1c5d21f86bbf62e18@%3Cdev.tinkerpop.apache.org%3E

Now that we have native python tests and the GLV test suite is almot done we can probably just rely on that for testing and not worry about this stuff. But at least the project builds now. CTR