fix race condition in TinkerIndex
My colleage @fabsx00 discovered a race condition in tinkergraph's index creation. He fixed it by simply replacing parallelStream with stream. Quoting his analysis:
So, reading the code, you see that this.put is called in parallel, but that method seems to contain a race as get is called on the index, checked for null, and a subsequent write is performed. It still seems like using stream here fixes the problem we've been seeing, and the performance hit is not significant.

Ticket: https://issues.apache.org/jira/browse/TINKERPOP-1830