Cleaned up the python build.
Docs were broken as a result of the jython maven plugin being disabled (it was disabled a while ago because it no longer works given authentication changes in pypi and a new plugin has not yet been released to support it). Removed some dead file coping which was no longer relevant. Included import of aenum so that docs would work - didn't seem necessary for compile. CTR