Improved the method for dealing with the python build problem with radish
The issue is definitely related to the colorful version bump with radish (and PyHamcrest, maybe others??) depend on. Rather than exclude the dependencies and manually import stuff which was messy, David Brown suggested the fix that just involved an environment variable. He is currently looking into opening an issue with colorful for the problem. CTR