Pinned netty to 4.0.x for spark/console
Spark and Gremlin Console use netty 4.0.x. They have to be the same version because spark is incompatible with 4.1.x so if the console has that version present then spark jobs won't submit and docs won't generate. Luckily, the driver is fine with 4.0.x in terms of connecting to 4.1.x from the testing i've tried.