TINKERPOP-1143 Moved TraversalEngine to gremlin-test
TraversalEngine and related infrastructure was deprecated way back on 3.2.0 but it was kept around because it was used in some infrastructure for testing. Graph providers should in no way still be using this interface and should have long ago gone to Computer as the gremlin-core infrastructure long ago quit using TraversalEngine in any capacity.