TINKERPOP-1705 Removed rebindings API from java driver
The "rebindings" API methods were long ago deprecated in 3.1.0-incubating but were not removed in 3.2.x or 3.3.x for extended support of that feature - especially for third-party drivers. By now, most users should have moved on to use the "aliases" API instead and most of the drivers should be supporting that as well, especially since most of the drivers are now officially TinkerPop maintained.