TINKERPOP-1869 Allowed iterate() after profile()
The iterate() method adds a NoneStep that was preventing traversals like g.V().profile().iterate(). Not the typical type of traversal you would do, but it seems like it should be allowed for consistency sake. Changed error message a bit as the user typically wouldn't add the "requirements step" manually so such a message would just be confusing.