Bumped build plugin versions to latest.
True of all plugins identified by the versions-maven-plugin except for the dotnet-maven-plugin which fails the build. Not sure why it fails, but I'm guessing that it has to do with some incompatibilty with dotnet command line or something. No longer need hyracs repo for anything so removed that entry from gremlin-groovy CTR