Remove deprecation from eval() method
This method was never meant to be deprecated. It seems to have been deprecated as a result of a bad merge. Looking back at tp32 it's clear that a previously removed method on this tp33 branch was the original one to be deprecated and this one was meant to stay. CTR