Fixed a bug in the CoreImports for pageRank() and peerPressure()
Docs weren't generating as a result of the imports. Tehcnically we needed a static class import which CoreImports didn't support so rather than try to re-work all that, I just pulled PageRank and PeerPressure into their own top-level classes so they would just work with what we had. Also, lower-cased the keys as that looked more consistent with Gremlin. CTR