TINKERPOP-1999 Prevented java driver from hanging on disorderly server shutdown
We didn't see this before because all tests had assumed an orderly shutdown. In orderly shutdown the server actually manages to send a kill message which the driver respects. In disorderly shutdown, the channel just goes inactive and as described in TINKERPOP-1999 that event was not being handled properly on the driver causing it to hang. I tried to write a test for this but was unable to as I can't figure out how to write an "disorderly shutdown" which basically involves "pulling the plug" on the server. CTR