TINKERPOP-1999 Fixed bug introduced in previous fix to this issue
The problem was introduced at 3dcabd4f1b23fecde28c79a10502ad562d934c8d - by throwing an exception on the client in channelInactive() it would trigger double reporting of exception in normal shutdowns and other odd behavior including errors in doc generation. Instead of an exception we just release pending messages which releases any blocked threads (which was the original nature of the problem this issue set out to fix). CTR