TINKERPOP-1972 Fix arg handling for inject step
To be precise, this affects all steps that have a params array with a
generic type parameter as the member type but currently inject is the
only step in Gremlin.Net that has such an array as an argument.

The arguments passed as such a params array were incorrectly wrapped in
one array before instead of individual arguments.