TINKERPOP-2084 Display remote stacktrace in gremlin console
Making the remote stacktrace visible to the user saves a visit to the server to determine what went amiss with failed script when the remote failing message is not enough to debug the problem. Plus, it makes the Gremlin Console in ":remote console" mode behave more like it is in local mode.