TINKERPOP-1984 Enable GraphSON serializers to no longer conflict
Making GraphSONMessageSerializerV2d0 handle both typed and untyped formats allows it to be used in a fashion where users can choose to add both serializers at the same time for HTTP use cases. It also generally seemed to make more sense than how it was before. Not sure what changed to allow this now and not in the past.