Add request pipelining TINKERPOP-1775/1774
This allows connections to be reused for different requests in parallel
which should lead to better utilization of connections / lower resource
usage. The ConnectionPool now also has a fixed size whereas it could
previously create an unlimited number of connections.
Each Connection can handle up to MaxInProcessPerConnection requests in
parallel.
If this limit is reached for all connections, then a
NoConnectionAvailableException is thrown which makes this a breaking
change.