Added contributions link in downloads
Replaces the javadoc link which seems kinda silly now with all the GLVs also having similar API links. the javadocs is also accessible immediately on the "documentation" link. Still need to do this for all the other releases on the download page but at least the pattern is there. CTR