TINKERPOP-2165 Preferred commons-lang3 to commons-lang
commons-lang is still part of TinkerPop because of commons-configuration and spark, but at least our internally used code doesn't have to be dependent on it. Hopefully we can utterly remove that unmaintained line at some point in the future. NOTICE no longer needs the commons-lang3 snippet as that aspect of their NOTICE has been removed.