TINKERPOP-2135/2148 Use custom COW collection
A Custom copy-on-write (COW) collection is used to store the connections
instead of a ConcurrentBag. This collection is heavily inspired by the
CopyOnWriteList in the DataStax C# Driver for Apache Cassandra.
This collection allows for an efficient removal of closed connections
which fixes TINKERPOP-2135. Some other operations like Count or the
removal of all connections are also more efficient.

This new CopyOnWriteCollection also enables round-robin scheduling of
requests over connections instead of searching for the least used
connection which required iterating over all connections and led to
problems when multiple threads in parallel got the same connection and
reached the max in process limit for that connection because of that.
So, this also fixes TINKERPOP-2148.

InternalsVisibleTo is added which allows us to test internal types like
the new CopyOnWriteCollection. Right now, only this collection is tested
like this, but we can add unit tests for other internal types in the
future, too.