TINKERPOP-2135/2148 Remove while loop to populate pool
Instead of implicitly waiting for another thread to populate the pool,
now the pool population is aborted immediately which leads to an
exception. The idea here is to fail fast instead to communicate the
problem to the user.

Also a small refactoring where an out parameter is turned into a return
value.