Proposed typo fix and minor grammatical stuff
I didn't think "heterogenous" was a word, but I looked it up, and it is!  :-)  But "heterogeneous" was intended here, right?

In the Community Contributions section, I think either of the constructions

     represent both [this] and [that]
or
     represent [this] as well as [that]

would be fine (but not "represent both [this] as well as [that]").

Are the proposed added hyphens in "Cypher to Gremlin traversal" -> "Cypher-to-Gremlin traversal" okay?