Typo fix to tp33
Did I target this fix to the correct branch?  A notification email I received containing a message from Stephen Mallette re my first pull request mentioned that 3.3.x is currently maintained on the tp33 branch, 3.4.x on tp34, and 3.5.0 (unreleased) on master, and said it'd be nice if I could target future PRs to the earliest branch to show the problem, so I checked the tp33 branch to see if the typo was there, and it was.  But I didn't check for anything earlier (tp32?) ... should I have?  Or is tp33 the earliest one to check?  Thanks!