TINKERPOP-2248 Force replacement of connections on certain errors
Relying only on the Connection.isDead() check to determine connection replacement seemed to introduce a regression as of 2cd84ff1f3944d67d44cbba5bb032a1c57377975 The shouldBlockRequestWhenTooBig was failing somewhat randomly since that change. Not sure if there is a better way to make isDead() be the only check needed for connection replacement in this context, but this change brings the logic back that checks exception types that were forcing replacement prior to thise change.