TINKERPOP-2272 Renamed steps/tokens conflicting with global function names
Used the standard underscore suffix on these problems. Didn't document a bunch purposely because the intention is to retain the old conflicting steps/tokens for the time being (both are currently present in the library so as to be backward compatible in 3.3.x and 3.4.x). In 3.5.0 all tests will be modified, documentation updated and the old conflicting names completely removed.