Explicitly closed traversal on interruption.
This seems to make sense even though it wasn't causing problems really. A traversal might not release resources on its own without an explicit close if the traversal does not complete iteration. Since we have an interruption of the traversal, an explicit close() seems like the right move here. CTR