TINKERPOP-2291 Added GraphSON support for deserialization of TraversalExplanation
Added the notion of ImmutableExplanation which allows the deserialization to occur. As an immutable object the recalculation of the TraversalExplanation does not need to occur from the raw data. In fact, if it did it would be "wrong" as it would recalculate on the client rather than the server and server side strategies may not be applied. That said, the design choices here in the class hierarchy are a bit odd in that ImmutableExplanation should really extend AbstractExplanation, but to avoid breaking changes in derser in gryo/graphson it seemed best to go with the slightly less optimal approach. Perhaps a future release branch were breaking changes are allowed can rectify this.