TINKERPOP-2284 Added elementMap() step
Given that we only return reference objects for graph elements now, there is a lot of extra project()/valueMap() work to do for the simple and common need to return all the structure of a graph element. elementMap() simplifies all that to a single step without a ton of options and added features.