Restored check of GSSException
Different environments seem to throw that exception from time to time. Not completely clear which ones do and which don't but ever since we started checking those we get much more consistent test runs. Probably need to better trap and unify exceptions in the driver but that's a different story/ticket/jira/issue/whatever CTR