Added back some tests that were inadvertantly removed from python
Not sure how they got removed - I think it was a rage delete at one point in trying to get this all to work. Added it back and uncovered a really bad bug in Edge deserialization which is now fixed. Side-effects are now working even though we've technically deprecated those now and even removed support along 3.5.0.