TINKERPOP-2289 Disabled IPv6 in tests
This change was a bit of a follow on to this issue where it was raised that INetAddress.getAllByName(String) might return both the IPv6 and IPv4 addresses when specifying a hostname to the cluster. That would mean that the Cluster would think that there were two hosts despite them likely pointing to the same instances (as in, localhost). By disabling IPv6 in tests we should avoid that problem for developers. Also added some javadoc to point folks in the right direction in case they run into such problems. CTR