TINKERPOP-2309 Bump to Tornado 5.x
Removed a test that started failing after the version bump. I assume it no longer is necessary given that the error is "Cannot run the event loop while another loop is running" and I've taken to mean that since asyncio is automatically use when available the test is no longer relevant.