TINKERPOP-2235 Fixed ImmutablePath again
Can't use null for the currentObject - has to be some kind of special indicator. Still not convinced that this is the best way to make this work, but so much of our infrastructure is predicated on null having some special meaning rather than having some property that describes its state.