TINKERPOP-2317 Removed Jython support
This was deprecated in 3.3.10 and has little value at this point given the direction we've gone with GLVs. We can now focus gremlin-python on being a python on library supporting the latest version of python only. Build speed improves dramatically. Lambdas are further minimized. No worries about the suboptimal and under-featured scriptengine implementation.