Reverted a portion of TINKERPOP-2325
Docs were not generating and we're close to release. Added a test that passes without the changes but fails completely with the changes. There are still improvements to make here, but I wonder how many of them should be made with match() strategies rather than sparql optimizations. if we fix up match() to improve normal index hits then we improve native match() usage and sparql. CTR