TINKERPOP-2336 Added getMaxWaitForClose java driver setting.
Deprecated the "close" message for sessions along with its driver setting. The driver should stay compatible with older server versions, but the server now closes sessions with the close of the connection and there no longer is a block while all messages return from the server after a Client.close().