Fixed bug in session based request tests/docs
Session based requests were using an actual UUID type for the session identifier. It's actually just supposed to be a string, but we often use UUID strings for convenience which seems to have led to the confusion. Unfortunately this issue has been there for a long time (since IO testing started). Opted to update the past versions and thankfully it didn't seem to create any ill effect. CTR