Manually installed python 3.5.3 for docker build
There's a problem with python 3.5.2 and earlier when it comes to some dependencies on the master branch (3.5.0 at this point). Something related to https://github.com/python/typing/issues/259 and of course xenial only pulls in 3.5.2. We stick to xenial because of 3.3.x/3.4.x...i guess we'd have to do equally oddball things to the docker image if we went to bionic so this change seems like the one of least effort. CTR