TINKERPOP-2345 Improved error message for bad value for math()
This problem really only applies to 3.4.x since by() didn't modulate Map traversers in 3.3.x. This change does also however fix the error message for non-Number objects trying to be passed into a math() expression, but the error is at least more understandable than the NullPointerException - on 3.3.x we get "java.lang.String cannot be cast to java.lang.Number" which while not perfect is probably good enough for that older version. It didn't seem like a big enough problem to add the enhancement there.