Added in .net session test improvement that were added in 3.3-dev
The change was merged forward from 3.3-dev but the change was ignored since it came with other baggage that didn't need to affect 3.4-dev and master. More specifically, the session feature was added to 3.4-dev first unfortunately and then later added on 3.3-dev after the fact and the 3.3-dev changes essentially conflict pretty hard with 3.4-dev at this point that it was just better to ignore them on merging the commits forward. CTR