Fixed doc generation in docker.
Grape failing in that environment for some reason, but the addition of the dual setting seemed to let the combined resolvers work. Fixed some issues with the docker image itself which seemed to be missing some commands that the doc gen has always relied on. Hard to believe that has been a problem since going to xenial but perhaps it is so. CTR