TINKERPOP-2365 Prevented extra barrier being added with profile()
LazyBarrierStrategy was improperly adding a barrier step to the end of a traversal when profile() was called. This wasn't noticeable in testing because of an overly zealous choice to disable the strategy in testing if profile() was present. Modified the strategy to more properly handle this situation and to limit the removal of LazyBarrierStrategy to ProfileTest only.