TINKERPOP-2421 Refactoring connection/client interaction.
There was a fair bit of tangled logic here especially after the adding of the notion of requestOptions. Tried to keep logic for request args out of the connection and make that just responsible for RequestMessage construction and serialization. Even though changes to Connection.submit() parameters were made I don't think it matters if we break this API here as users shouldn't be coding at this level - they will be either using Client to send scripts or using bytecode based requests with remote traversals.