TINKERPOP-2443 Refactored TypeTranslator
Tried to make better re-use of the logic of the TypeTranslator. There was a lot of duplication between python and groovy and it seemed like that could be abstracted a bit so that the language specific implementations could focus more solely on the actual types that needed to be translated.