TINKERPOP-2443 Preserve variables in Gremlin string translation
Fixed a little gang of bugs, include TraversalStrategy issues in Python and TINKERPOP-2458 with this commit which will make sure bytecode Bindings aren't lost in some cases. Some of this work should be backported to 3.4.x. Had to remove a bunch of tests to get the core of this body of work operational - intend to add them back on a future commit.