TINKERPOP-2469 Improved exception handling out of KdcFixture
When the KdcFixture fails with an exception in test teardown it tanks the test that was part of that generating a false positive for a failure. This change doesn't quite get to the root cause of the problem, at least not insomuch as I can discern, but I sense that a missing principal during teardown should not be evidence of failure especially if the core of the tests passed in the first place. Would be nice to know how the principal is getting remove (or is not created in the first place while still allowing the tests to pass) but I think the key to get Travis stable for this change. CTR