TINKERPOP-2438 Improved performance of GremlinASTChecker
Used CompileStatic annotation and introduced a lightweight check to validate that there was reason to accept the expense of parsing. It's a bit of a hack in a sense but after some analysis it seems that there are no easy fixes. Everything ends in fairly advanced refactoring and extension of GremlinGroovyScriptEngine and GremlinGroovyClassLoader (with possible pull requests to Groovy). That seems like a fair bit of effort for what this feature is and I'm not sure it's worth it. The typical use case for timeout setting should not encounter the performance problem that would ensue from parsing to the Groovy AST.