TINKERPOP-2472 Decouple the driver from the IO format
It is now easier to use the driver with a serialization format that is
not GraphSON or even JSON based. This should allow us to support
GraphBinary.

NOTE: The driver now always deserializes a result completely, e.g.,
skipping serialization like it was implemented in TINKERPOP-2067 is no
longer supported. However, users who need this functionality can simply
implement their own IMessageSerializer with this functionality.