TINKERPOP-2481 Installed IdentityRemovalStrategy
This strategy was always present but never installed. Research for the reasoning is written up on the JIRA for this issue. Needed to change the Step interface to allow for a way to access start objects without iterating them to get rid of a bit of hackery in BranchStep which needed access to such information without actually processing its child traversals.