Removed line discouraging scripts.
Not sure we need to say this anymore as we can't realistically get rid of scripts unless we have some feature parity with bytecode and we can't do that without a solution for remote transactions. There is also a push toward the liklihood of gremlin-script which would rid us of script security issues. In short, ideas/direction are shifting in this area. CTR