TINKERPOP-2472 Support GraphBinary in Gremlin.Net
Still missing:
* docs
* metric serializers

Executing the Gherkin tests with GraphSON and GraphBinary is
unfortunately not easy to implement as a lot of the graph initialization
and connection creation logic in our Gherkin implementation is static.
So this would require a lot of work to become more flexible. Given
that we already have a lot of GraphBinary tests, I think that it's OK
for now to not run all tests with Gherkin, although it would definitely
be nice to have.

GraphBinary is faster for all test cases than GraphSON 3. It
unfortunately allocates more memory during deserialization.

Detailed results:

``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19041.804 (2004/?/20H1)
Intel Core i7-10850H CPU 2.70GHz, 1 CPU, 12 logical and 6 physical cores
.NET Core SDK=5.0.100
  [Host]     : .NET Core 3.1.9 (CoreCLR 4.700.20.47201, CoreFX 4.700.20.47203), X64 RyuJIT
  Job-JJMOFB : .NET Core 3.1.9 (CoreCLR 4.700.20.47201, CoreFX 4.700.20.47203), X64 RyuJIT

IterationCount=5  LaunchCount=1  WarmupCount=3

```
|                                Method |       Mean |      Error |    StdDev |    Gen 0 |   Gen 1 | Gen 2 | Allocated |
|-------------------------------------- |-----------:|-----------:|----------:|---------:|--------:|------:|----------:|
|          TestWriteEmptyBytecodeBinary |   4.114 μs |  0.1149 μs | 0.0298 μs |   0.3357 |       - |     - |    2.1 KB |
|       TestWriteEmptyBytecodeGraphSON3 |   5.503 μs |  0.1823 μs | 0.0473 μs |   0.5417 |       - |     - |   3.33 KB |
|               TestWriteBytecodeBinary |  22.241 μs |  0.3235 μs | 0.0840 μs |   1.3733 |       - |     - |   8.45 KB |
|            TestWriteBytecodeGraphSON3 |  35.849 μs |  0.6938 μs | 0.1074 μs |   2.9297 |       - |     - |  18.27 KB |
|    TestReadSmallResponseMessageBinary |   5.706 μs |  0.3858 μs | 0.0597 μs |   1.4954 |  0.0076 |     - |   9.19 KB |
| TestReadSmallResponseMessageGraphSON3 |  10.025 μs |  0.9566 μs | 0.2484 μs |   0.5493 |       - |     - |   3.39 KB |
|      TestReadBigResponseMessageBinary | 622.829 μs | 10.1589 μs | 1.5721 μs | 140.6250 | 34.1797 |     - | 865.06 KB |
|   TestReadBigResponseMessageGraphSON3 | 776.591 μs | 28.7843 μs | 4.4544 μs |  36.1328 | 11.7188 |     - | 224.24 KB |