TINKERPOP-2472 Gherkin now also tests GraphBinary
I also tried adding GraphSON 2 for Gherkin but that lets a bunch of
tests fail which assert on things like T.label that were simply returned
as a string "label" in GraphSON 2.
So these would need some more work which doesn't seem worth it for a
message format that we deprecated years ago.