TINKERPOP-2245 Refactored the session side of UnifiedChannelizer
MultiRexster didn't have quite the same semantics as the SessionOpProcessor. Had to add in an option to configure the MultiRexster to work that way. Also sorted out some issues with timeouts with sessions which required a RexsterExecutorService so that a Thread processing a Rexster could be interrupted independently of the Future as calling cancel() on the Future is something that can only be done once and if you wish to preserve state between failures of requests in a session (i.e. the OpProcessor way) then the interruption can't be final.