TINKERPOP-2245 gremlinPool needs to be defaulted to a larger value
For environments like travis which might only have two cores there is a chance for tests to fail/not finish as they introduce blocks in requests that need to exist for tests to pass. If there aren't enough threads in the pool the requests can't be serviced properly and they fail.