TINKERPOP-2569 Reconnect if java driver fails to initialize
Does result in a slight change in behavior as client initialization is now aware of dead hosts and will throw a NoHostAvailableException earlier (used to wait for a request to be submitted). This is probably better behavior. It seems unlikely that anyone would rely on this behavior for normal usage so pushing it to 3.4.x seems safe.