Generate updated package-lock.json for gremlin-javascript CTR
The current package-lock.json file in gremlin-javascript was merged
from 3.4-dev, which means it was generated based on the package.json
file in 3.4-dev. The package.json file in 3.5-dev is not identical, so
this commit makes sure the package-lock.json file is correctly derived
from the package.json file.

When the package-lock.json file conflicts with the package.json file,
a new package-lock.json file will be generated when running npm
install. This is not a big deal, but it can be annoying to get an
unstaged change that will have to be stashed or committed whenever
you run mvn clean install (which runs npm install behind the scenes).