TINKERPOP-2358 Stop leaking connections on Dispose
If the `ConnectionPool` was disposed while it was in parallel creating
new connections (e.g., to replace closed connections), those
connections could be leaked. `Dispose()` could not dispose them yet as
they were not completely established so they could be added to the pool.

We now check after creating new connections whether the pool has been
disposed in the meantime and then dispose these connections directly
again.

In addition to that, the pool now also has a `CancellationTokenSource`
which allows us to cancel all active creations of new connections in
`Dispose()`. So we don't have to wait until they are created only so we
can then dispose them if we can already cancel the connection
establishment.