TINKERPOP-2507 Allowed graph filters with mixed id types
Upgrade docs in this commit explain this change fairly well. There doesn't seem to be a need to be so prescriptive about the types that can be filtered on and it should be up to each graph to decide how they handle such input. Removed the general exception, tests and updated the behavior of graphs to support the change.