TINKERPOP-2576 Enable GitHub's Dependabot for .NET
Dependabot is now part of GitHub which comes with a new config file.
This commit configures Dependabot only for Gremlin.NET so we can try it
out there first.

We don't need the old Dependabot config file any more that was for the
pre-GitHub version as soon as we enable Dependabot via GitHub by merging
this PR.